requirements.txt

fastapi
uvicorn
sqlalchemy
pyyaml
pydantic
